before change

+++ 0_proj/3Dsim/3Dsim/initialize.h
@@ -481,7 +481,7 @@ struct request{
 	unsigned int size;                 //The size of the request, the number of sectors
 	unsigned int operation;            //The type of request, 1 for the read, 0 for the write
 	unsigned int cmplt_flag;		   //Whether the request is executed, 0 means no execution, 1 means it has been executed
-	double compressed_size;
+	unsigned int compressed_size;
 
 	unsigned int* need_distr_flag;
 	unsigned int complete_lsn_count;   //record the count of lsn served by buffer
